Detecting user clicking Cancel on Parameter editor dialog and abort the execution.
